common: Split out the deserialized responses
